Re-organize calendar service directory; add launch.json
