+++ src/utils/__tests__/mapEventStatus.tests.ts
@@ -1,5 +1,5 @@
 import { getStatusForUserEvent } from '../mapEventStatus';
-import { ShowAs } from '../../services/calendar/calendar';
+import { ShowAs } from '../../services/calendar';
 
 const baseUserSettings = { email: 'test@email.com', slackToken: 'abc' };
